R2Gen-Finetuning-main/requirements.txt
alembic==1.16.1
beautifulsoup4==4.13.4
bert-score==0.3.13
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.5.0
gdown==5.2.0
greenlet==3.2.2
huggingface-hub==0.32.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.6
opencv-python==4.11.0.86
optuna==4.3.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pycocoevalcap==1.2
pycocotools==2.0.8
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
setuptools==80.8.0
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.41
sympy==1.13.1
tokenizers==0.21.1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.52.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
